refactor: Refactor elements in projects and themes pages for better server-side rendering Additonally created custom hook to get nav drawer list
